camera_side/requirements.txt
picamera
Pillow
ConfigParser
argparse
